Add an option `doxygen-bin-path` to specify the path to `doxygen`.
This commit introduces an option to let the user specify the filepath to the
`doxygen` binary.

This option is called `doxygen-bin-path` and can be set in the `mkdocs.yml` under
the `mkdoxy` section.

The following snippet is an example of how to use this new option:

```yaml
site_name: My Doc
…

plugins:
  - mkdoxy:
    enabled: !ENV [ENABLE_MKDOXY, True]
    doxygen-bin-path: path/to/doxygen/build/bin/doxygen
    projects:
      mkdoxyApi:
        src-dirs: mkdoxy
        …
```

If not specified, the default value is obviously `doxygen`.